pyproject.toml
[tool.poetry]
name = "password-manager-api"
version = "0.1.0"
description = ""
authors = ["Dmitrii Morozov <33morozov@gmail.com>"]
readme = "README.md"
packages = [{include = "password_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
flake8 = "^6.0.0"
sqlmodel = "^0.0.8"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
